// FootballManagement/football-ng/src/app/Model/match-detail.ts
export class MatchDetail {
   id: number;
   happenedDate: Date;
   playInMatchId: number;
   playerId: number;
   playerName: string;
   teamId: number;
   teamName: string;
   cardReceived: string;
   action_time: number;
   is_goal_from_11m: boolean;
   is_the_bad_goal: boolean;
   is_Goal: boolean;
   is_MoveIn: boolean;
   is_MoveOut: boolean;
}